Adds useEmbeddedWalletSendVerificationEmail
