+++ packages/react-native/src/evm/index.ts
@@ -32,6 +32,7 @@ export { useRainbowWallet } from "./wallets/hooks/useRainbowWallet";
 export { useTrustWallet } from "./wallets/hooks/useTrustWallet";
 export { useEmbeddedWallet } from "./wallets/hooks/useEmbeddedWallet";
 export { useSmartWallet } from "./wallets/hooks/useSmartWallet";
+export { useEmbeddedWalletSendVerificationEmail } from "./wallets/hooks/useEmbeddedWalletSendVerificationEmail";
 
 export {
   ConnectWallet,
