+++ packages/react-native/src/evm/wallets/hooks/useSmartWallet.ts
@@ -5,6 +5,7 @@ import {
   useSetConnectedWallet,
   useSetConnectionStatus,
   useWalletContext,
+  useWallets,
 } from "@thirdweb-dev/react-core";
 import {
   getSmartWalletAddress,
@@ -13,7 +14,6 @@ import {
 import { useCallback } from "react";
 import type { BytesLike } from "ethers";
 import { SmartWalletConfig } from "../types/smart-wallet";
-import { EmbeddedWallet } from "../wallets/embedded/EmbeddedWallet";
 
 export function useSmartWallet<W extends WalletInstance>(
   personalWallet: WalletConfig<W>,
@@ -23,6 +23,7 @@ export function useSmartWallet<W extends WalletInstance>(
   const setStatus = useSetConnectionStatus();
   const setWallet = useSetConnectedWallet();
   const context = useWalletContext();
+  const supportedWallets = useWallets();
 
   const predictAddress = useCallback(
     async (args: { personalWalletAddress: string; data?: BytesLike }) => {
@@ -41,6 +42,11 @@ export function useSmartWallet<W extends WalletInstance>(
       connectPersonalWallet?: (wallet: W) => Promise<void>;
       connectionArgs?: Omit<SmartWalletConnectionArgs, "personalWallet">;
     }) => {
+      if (!supportedWallets.find((w) => w.id === personalWallet.id)) {
+        console.warn(
+          "Please, add your smart wallet to the supportedWallets prop of the ThirdwebProvider to enjoy auto-connecting to the wallet.",
+        );
+      }
       const { smartWallet } = await import("../wallets/smart-wallet");
       setStatus("connecting");
       const pw = create(personalWallet);
@@ -75,25 +81,12 @@ export function useSmartWallet<W extends WalletInstance>(
       personalWallet,
       setStatus,
       setWallet,
+      supportedWallets,
     ],
-  );
-
-  const sendVerificationEmail = useCallback(
-    async (email: string) => {
-      if (!context.clientId) {
-        throw new Error("Please, provide a clientId in the ThirdwebProvider");
-      }
-      return EmbeddedWallet.sendVerificationEmail({
-        email,
-        clientId: context.clientId,
-      });
-    },
-    [context.clientId],
   );
 
   return {
     connect,
     predictAddress,
-    sendVerificationEmail,
   };
 }
